:art: Code cleanup (#666)
* :green_heart: Remove unused help jar

* :art: Auto format code

* :art: Organise imports

* :art: Inspect and refactor (NetBeans option)
* Convert to switch over Strings
* Remove .floatValue()
* Remove .intValue()
* Convert to for-loop (skipped some)
* Some lambda expression (skipped some)
* Repalce with multicatch
* Use member reference (skipped some)
* Convert to try with resoruce

* :race_horse: Performance Improvements
Used the NetBeans Inspect and Refactor option and ran through all possible performance options and pretty much implemented them all.

* :art: Remove unused imports

+++ CoreGraphFramework/src/au/gov/asd/tac/constellation/graph/attribute/ByteObjectAttributeDescription.java
@@ -111,7 +111,7 @@ public class ByteObjectAttributeDescription extends AbstractObjectAttributeDescr
 
     @Override
     public double getDouble(final int id) {
-        return data[id] != null ? ((Byte) data[id]).byteValue() : 0.0;
+        return data[id] != null ? ((Byte) data[id]) : 0.0;
     }
 
     @Override
